+++ src/MetricKind.php
@@ -11,5 +11,6 @@ enum MetricKind: string
 {
     case Counter = 'counter';
     case Gauge = 'gauge';
+    case UpDownCounter = 'up_down_counter';
     case Histogram = 'histogram';
 }